+++ client/app/shared/modals/upload-image-modal/upload-image-modal.html
@@ -12,7 +12,10 @@
 
         <md-dialog-content>
             <div class="md-dialog-content">
-                <photo-upload id="uploadPhoto" ng-model="photo" area-type="circle"/>
+                <photo-upload id="uploadPhoto"
+                    ng-model="photo"
+                    accept-type="image/png, image/jpeg"
+                    area-type="circle"/>
             </div>
         </md-dialog-content>
